chore(query): extracting processes related to http clients in `query`
